refactor: one colour-scheme composable for mail and calendar
The theme is a single User Settings row, but each app kept its own copy of
the reader, and the calendar's cycle shortcut had none of mail's write
queue — so a fast Cmd+Shift+L could send two set_value calls against one
`modified` and fail the second. Both apps now go through a suite-wide
useTheme that takes the app's user resource.
